observability-stack-demo/app/requirements.txt
fastapi
uvicorn[standard]
prometheus-client
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-logging
python-json-logger
